Cleanup and an IAPService fix
