+++ CatchUp-SwiftUI/HomeScreen.swift
@@ -18,15 +18,9 @@ struct HomeScreen : View {
     @AppStorage("savedVersion") var savedVersion = "2.0.0"
 
     @State private var isColdLaunch = true
-    @State private var isShowingContactPicker = false
 	@State private var isShowingUpdatesSheet = false
     @State private var isShowingAboutSheet = false
-
     @State private var contactPicker = ContactPicker()
-
-    let notificationService = NotificationService()
-    let helper = GeneralHelpers()
-	let converter = Conversions()
 
 	init() {
         //Use this if NavigationBarTitle is with Large Font
@@ -40,7 +34,7 @@ struct HomeScreen : View {
 					ForEach(selectedContacts) { contact in
 						NavigationLink(destination: DetailScreen(contact: contact)) {
 							HStack {
-                                converter.getContactPicture(from: contact.picture)
+                                Converter.getContactPicture(from: contact.picture)
 									.renderingMode(.original)
 									.resizable()
 									.frame(width: 45, height: 45, alignment: .leading)
@@ -49,7 +43,7 @@ struct HomeScreen : View {
 								VStack(alignment: .leading, spacing: 2) {
 									Text(contact.name)
 										.font(.headline)
-									Text(converter.convertNotificationPreferenceIntToString(preference: Int(contact.notification_preference), contact: contact))
+									Text(Converter.convertNotificationPreferenceIntToString(preference: Int(contact.notification_preference), contact: contact))
 										.font(.caption)
 										.foregroundColor(.gray)
 								}
@@ -65,9 +59,9 @@ struct HomeScreen : View {
                     contactPicker.chosenContacts = []
                 }
 
-                Button(action: {
+                Button {
                     openContactPicker()
-                }) {
+                } label: {
 					HStack(alignment: .center, spacing: 6) {
 						Image(systemName: "person.crop.circle.fill.badge.plus")
 						
@@ -78,14 +72,6 @@ struct HomeScreen : View {
 					.padding(.top)
 					.padding(.bottom)
                 }
-//                .sheet(isPresented: $isShowingContactPicker) {
-//                    ContactPicker(
-//                        showPicker: $isShowingContactPicker,
-//                        onSelectContacts: { selectedContacts in
-//                            saveSelectedContact(for: selectedContacts)
-//                        }
-//                    )
-//                }
 
 				.navigationBarTitle(Text("CatchUp"))
 					
@@ -106,48 +92,14 @@ struct HomeScreen : View {
 		.accentColor(.orange)
 
         .onAppear {
-            print("HomeScreen onAppear")
+            print("sqlite path: \(modelContext.sqliteCommand)")
             clearNotificationBadgeAndCheckForUpdate()
-            notificationService.requestAuthorizationForNotifications()
+            NotificationHelper.requestAuthorizationForNotifications()
 
             if isColdLaunch {
-                resetNotifications()
+                NotificationHelper.resetNotifications(for: selectedContacts, modelContext: modelContext)
                 isColdLaunch = false
             }
-
-            print(modelContext.sqliteCommand)
-        }
-    }
-    
-    func clearNotificationBadgeAndCheckForUpdate() {
-		fetchAvailableIAPs()
-        helper.clearNotificationBadge()
-        try? modelContext.save()
-        checkForUpdate()
-    }
-    
-    func removePendingNotificationsAndDeleteContact(at offsets: IndexSet) {
-        for index in offsets {
-            let contact = selectedContacts[index]
-            
-            notificationService.removeExistingNotifications(for: contact)
-            modelContext.delete(contact)
-        }
-    }
-    
-    func checkForUpdate() {
-        let version = helper.getCurrentAppVersion()
-        print("latest version: \(version)")
-
-        if savedVersion == version {
-            print("App is up to date!")
-        } else {
-			if updateIsMajor() {
-				// Toggle to show UpdatesScreen as a sheet
-				print("Major update detected, showing UpdatesScreen...")
-				isShowingUpdatesSheet = true
-			}
-            savedVersion = version
         }
     }
 
@@ -160,93 +112,16 @@ struct HomeScreen : View {
         window?.rootViewController?.present(contactPicker, animated: true, completion: nil)
     }
 
-    func resetNotifications() {
-        for contact in selectedContacts {
-            notificationService.removeExistingNotifications(for: contact)
-            notificationService.createNewNotification(for: contact, modelContext: modelContext)
-        }
-    }
-
-	func updateIsMajor() -> Bool {
-		let version = helper.getCurrentAppVersion()
-		if version.suffix(2) == ".0" {
-			return true
-		} else {
-			return false
-		}
-	}
-	
-	func fetchAvailableIAPs() {
-		print("fetching IAPs")
-		IAPService.shared.fetchAvailableProducts()
-	}
-
-    // save selected contacts and their properties to Core Data
+    // save selected contacts and their properties to SwiftData
     func saveSelectedContact(for contacts: [CNContact]) {
-        print("saving...")
-
-        let contactService = ContactService()
         for contact in contacts {
-            let currentMinute = Calendar.current.component(.minute, from: Date())
-            let currentHour = Calendar.current.component(.hour, from: Date())
-            let currentDay = Calendar.current.component(.day, from: Date())
-            let currentMonth = Calendar.current.component(.month, from: Date())
-            let currentYear = Calendar.current.component(.year, from: Date())
-
-            let id = UUID()
-            let address = contactService.getContactPrimaryAddress(for: contact)
-            let anniversary = contactService.getContactAnniversary(for: contact)
-            let anniversary_notification_ID = UUID()
-            let birthday = contactService.getContactBirthday(for: contact)
-            let birthday_notification_ID = UUID()
-            let email = contactService.getContactPrimaryEmail(for: contact)
-            let name = contactService.getContactName(for: contact)
-            let notification_identifier = UUID()
-            let notification_preference = 0
-            let notification_preference_hour = currentHour
-            let notification_preference_minute = currentMinute
-            let notification_preference_weekday = 0
-            let notification_preference_custom_year = currentYear
-            let notification_preference_custom_month = currentMonth
-            let notification_preference_custom_day = currentDay
-            let phone = contactService.getContactPrimaryPhone(for: contact)
-            let picture = contactService.encodeContactPicture(for: contact)
-            let secondary_email = contactService.getContactSecondaryEmail(for: contact)
-            let secondary_address = contactService.getContactSecondaryAddress(for: contact)
-            let secondary_phone = contactService.getContactSecondaryPhone(for: contact)
-
-            if !contactAlreadyAdded(name: name) {
-                let selectedContact = SelectedContact(
-                    address: address,
-                    anniversary: anniversary,
-                    anniversary_notification_id: anniversary_notification_ID,
-                    birthday: birthday,
-                    birthday_notification_id: birthday_notification_ID,
-                    email: email,
-                    id: id,
-                    name: name,
-                    notification_identifier: notification_identifier,
-                    notification_preference: notification_preference,
-                    notification_preference_custom_day: notification_preference_custom_day,
-                    notification_preference_custom_month: notification_preference_custom_month,
-                    notification_preference_custom_year: notification_preference_custom_year,
-                    notification_preference_hour: notification_preference_hour,
-                    notification_preference_minute: notification_preference_minute,
-                    notification_preference_weekday: notification_preference_weekday,
-                    phone: phone,
-                    picture: picture,
-                    secondary_address: secondary_address,
-                    secondary_email: secondary_email,
-                    secondary_phone: secondary_phone
-                )
-
+            let contactName = ContactHelper.getContactName(for: contact)
+            if !contactAlreadyAdded(name: contactName) {
+                let selectedContact = ContactHelper.createSelectedContact(contact: contact)
                 modelContext.insert(selectedContact)
-            } else {
-                print("Do nothing. Contact was already added to the database")
             }
         }
     }
-
 
     func contactAlreadyAdded(name: String) -> Bool {
         for contact in selectedContacts {
@@ -254,8 +129,39 @@ struct HomeScreen : View {
                 return true
             }
         }
-
         return false
+    }
+
+    func clearNotificationBadgeAndCheckForUpdate() {
+        Utils.fetchAvailableIAPs()
+        Utils.clearNotificationBadge()
+
+        checkForUpdate()
+    }
+    
+    func removePendingNotificationsAndDeleteContact(at offsets: IndexSet) {
+        for index in offsets {
+            let contact = selectedContacts[index]
+            
+            NotificationHelper.removeExistingNotifications(for: contact)
+            modelContext.delete(contact)
+        }
+    }
+    
+    func checkForUpdate() {
+        let version = Utils.getCurrentAppVersion()
+        print("latest version: \(version)")
+
+        if savedVersion == version {
+            print("App is up to date!")
+        } else {
+            if Utils.updateIsMajor() {
+				// Toggle to show UpdatesScreen as a sheet
+				print("Major update detected, showing UpdatesScreen...")
+				isShowingUpdatesSheet = true
+			}
+            savedVersion = version
+        }
     }
 }
