+++ CatchUp-SwiftUI/Services/IAPService.swift
@@ -59,27 +59,15 @@ class IAPService: NSObject {
     }
 	
 	func getSmallTipAmount() -> String {
-		if iapProducts.count > 0 {
-			return iapProducts[1].localizedPrice
-		} else {
-			return "$0.99"
-		}
+        return iapProducts.first(where: { $0.productIdentifier == "gracious_tip_0.99" })?.localizedPrice ?? "$0.99"
 	}
 	
 	func getMediumTipAmount() -> String {
-		if iapProducts.count > 0 {
-			return iapProducts[0].localizedPrice
-		} else {
-			return "$1.99"
-		}
+        return iapProducts.first(where: { $0.productIdentifier == "generous_tip_1.99" })?.localizedPrice ?? "$1.99"
 	}
 	
 	func getLargeTipAmount() -> String {
-		if iapProducts.count > 0 {
-			return iapProducts[2].localizedPrice
-		} else {
-			return "$4.99"
-		}
+        return iapProducts.first(where: { $0.productIdentifier == "gratuitous_tip_4.99" })?.localizedPrice ?? "$4.99"
 	}
     
 	// MARK: - RESTORE PURCHASE
@@ -103,7 +91,6 @@ class IAPService: NSObject {
 extension IAPService: SKProductsRequestDelegate, SKPaymentTransactionObserver{
     // MARK: - REQUEST IAP PRODUCTS
     func productsRequest (_ request:SKProductsRequest, didReceive response:SKProductsResponse) {
-        
         if (response.products.count > 0) {
             iapProducts = response.products
             for product in iapProducts{
